make it harder to trigger overflows
with just int that happens fast even for 'normal problems':

src/CbcModel.cpp:459:35: runtime error: signed integer overflow: 2074221009 + 91406250 cannot be represented in type 'int'

+++ src/CbcModel.cpp
@@ -427,7 +427,7 @@ void CbcModel::analyzeObjective()
     int numberInteger = 0;
     int numberIntegerObj = 0;
     int numberGeneralIntegerObj = 0;
-    int numberIntegerWeight = 0;
+    int64_t numberIntegerWeight = 0;
     int numberContinuousObj = 0;
     double cost = COIN_DBL_MAX;
     for (iColumn = 0; iColumn < numberColumns; iColumn++) {
